+++ contest/ClearController.java
