+++ contest/MapClear.fxml
